create minting service and test
