--- apps/Csharp.CardanoSounds/CS.TokenMint/IExamples.cs
@@ -1,7 +1,0 @@
-﻿namespace CS.TokenMint
-{
-    public interface IExamples
-    {
-        public void TestMintTokens();
-    }
-}
