+++ apps/Csharp.CardanoSounds/CS.TokenMint/IMint.cs
@@ -1,0 +1,7 @@
+﻿namespace CS.TokenMint
+{
+    public interface IMint
+    {
+        public void MintFromDbTransaction();
+    }
+}
